+++ patch_class_post_review.php
@@ -1,0 +1,2 @@
+<?php
+// We will use replace_with_git_merge_diff to add ajax_save_inline_edit_post to the controller.