+++ patch_post_review.php
@@ -1,0 +1,2 @@
+<?php
+// We will use replace_with_git_merge_diff to add inline editing fields to the rows.